ai-agent-projects/multi-ai-agent/guardrailed-tool-handoff-agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "guardrailed-tool-use-agent"
version = "0.1.0"
description = "Guardrailed multi-agent scaffold with handoffs, nested search-as-agent, and trace-aware sentiment reporting."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.13.6",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.12.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
